1.fix Graphql curl export variable body string issue 2.Clear reqeust body text when user delete both query&variable payload in GraphQL editor
