+++ packages/insomnia/src/network/unit-test-feature.ts
@@ -1,5 +1,6 @@
 import { stats } from '../models';
 import { getBodyBuffer } from '../models/response';
+import { parseGraphQLReqeustBody } from '../utils/graph-ql';
 import { fetchRequestData, responseTransform, sendCurlAndWriteTimeline, tryToInterpolateRequest, tryToTransformRequestWithPlugins } from './network';
 
 export function getSendRequestCallback() {
@@ -19,17 +20,7 @@ export function getSendRequestCallback() {
     const renderedRequest = await tryToTransformRequestWithPlugins(renderResult);
 
     // TODO: remove this temporary hack to support GraphQL variables in the request body properly
-    if (renderedRequest && renderedRequest.body?.text && renderedRequest.body?.mimeType === 'application/graphql') {
-      try {
-        const parsedBody = JSON.parse(renderedRequest.body.text);
-        if (typeof parsedBody.variables === 'string') {
-          parsedBody.variables = JSON.parse(parsedBody.variables);
-          renderedRequest.body.text = JSON.stringify(parsedBody, null, 2);
-        }
-      } catch (e) {
-        console.error('Failed to parse GraphQL variables', e);
-      }
-    }
+    parseGraphQLReqeustBody(renderedRequest);
 
     const response = await sendCurlAndWriteTimeline(
       renderedRequest,
